Uptade to BigInteger

+++ src/main/java/com/arbehr/labseq/util/LabseqHashMap.java
@@ -1,16 +1,17 @@
 package com.arbehr.labseq.util;
 
+import java.math.BigInteger;
 import java.util.HashMap;
 
 public class LabseqHashMap {
     
-    private final HashMap<Integer, Integer> labSeqHashMap;
+    private final HashMap<Integer, BigInteger> labSeqHashMap;
 
     public LabseqHashMap() {
         this.labSeqHashMap = new HashMap<>();
     }
 
-    public void insert(Integer n, Integer result) {
+    public void insert(Integer n, BigInteger result) {
         this.labSeqHashMap.put(n, result);
     }
 
@@ -18,7 +19,7 @@ public class LabseqHashMap {
         return this.labSeqHashMap.containsKey(n);
     }
 
-    public Integer getResult(Integer n) {
+    public BigInteger getResult(Integer n) {
         return this.labSeqHashMap.get(n);
     }
 }
